+++ java/client/src/org/openqa/selenium/chrome/ChromeDriver.java
@@ -19,6 +19,7 @@
 package org.openqa.selenium.chrome;
 
 import com.google.common.collect.ImmutableMap;
+
 import org.openqa.selenium.Capabilities;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebDriverException;
@@ -29,11 +30,13 @@ import org.openqa.selenium.html5.SessionStorage;
 import org.openqa.selenium.html5.WebStorage;
 import org.openqa.selenium.interactions.HasTouchScreen;
 import org.openqa.selenium.interactions.TouchScreen;
+import org.openqa.selenium.mobile.NetworkConnection;
 import org.openqa.selenium.remote.FileDetector;
 import org.openqa.selenium.remote.RemoteTouchScreen;
 import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.remote.html5.RemoteLocationContext;
 import org.openqa.selenium.remote.html5.RemoteWebStorage;
+import org.openqa.selenium.remote.mobile.RemoteNetworkConnection;
 
 /**
  * A {@link WebDriver} implementation that controls a Chrome browser running on the local machine.
@@ -104,11 +107,12 @@ import org.openqa.selenium.remote.html5.RemoteWebStorage;
  * @see ChromeDriverService#createDefaultService
  */
 public class ChromeDriver extends RemoteWebDriver
-    implements LocationContext, WebStorage, HasTouchScreen {
+    implements LocationContext, WebStorage, HasTouchScreen, NetworkConnection {
 
   private RemoteLocationContext locationContext;
   private RemoteWebStorage webStorage;
   private TouchScreen touchScreen;
+  private RemoteNetworkConnection networkConnection;
 
   /**
    * Creates a new ChromeDriver using the {@link ChromeDriverService#createDefaultService default}
@@ -175,6 +179,7 @@ public class ChromeDriver extends RemoteWebDriver
     locationContext = new RemoteLocationContext(getExecuteMethod());
     webStorage = new  RemoteWebStorage(getExecuteMethod());
     touchScreen = new RemoteTouchScreen(getExecuteMethod());
+    networkConnection = new RemoteNetworkConnection(getExecuteMethod());
   }
 
   @Override
@@ -208,7 +213,17 @@ public class ChromeDriver extends RemoteWebDriver
   public TouchScreen getTouch() {
     return touchScreen;
   }
-  
+
+  @Override
+  public ConnectionType getNetworkConnection() {
+    return networkConnection.getNetworkConnection();
+  }
+
+  @Override
+  public ConnectionType setNetworkConnection(ConnectionType type) {
+    return networkConnection.setNetworkConnection(type);
+  }
+
   /**
    * Launches Chrome app specified by id.
    *