+++ java/client/test/org/openqa/selenium/firefox/FirefoxSpecificTests.java
@@ -19,15 +19,11 @@ package org.openqa.selenium.firefox;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
-import org.openqa.selenium.firefox.internal.ExecutableTest;
 import org.openqa.selenium.firefox.internal.NewProfileExtensionConnectionTest;
 import org.openqa.selenium.firefox.internal.SocketLockTest;
-import org.openqa.selenium.firefox.internal.StreamsTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
-    CapabilitiesToProfileTest.class,
-    ExecutableTest.class,
     FirefoxCapabilitiesTest.class,
     FirefoxDriverTest.class,
     FirefoxProfileTest.class,
@@ -35,7 +31,6 @@ import org.openqa.selenium.firefox.internal.StreamsTest;
     NewProfileExtensionConnectionTest.class,
     PreferencesTest.class,
     SocketLockTest.class,
-    StreamsTest.class
 })
 
 public class FirefoxSpecificTests {