+++ java/client/test/org/openqa/selenium/firefox/MarionetteTest.java
@@ -17,17 +17,12 @@
 
 package org.openqa.selenium.firefox;
 
-import static org.junit.Assert.assertEquals;
 import static org.openqa.selenium.testing.Driver.FIREFOX;
-
-import com.google.common.base.Function;
 
 import org.junit.After;
 import org.junit.Test;
-import org.openqa.selenium.JavascriptExecutor;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
-import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.testing.Ignore;
 import org.openqa.selenium.testing.JUnit4TestBase;
 
@@ -44,6 +39,13 @@ public class MarionetteTest extends JUnit4TestBase {
   }
 
   @Test
+  public void canStartDriverWithNoParameters() throws InterruptedException {
+    driver = new FirefoxDriver();
+    driver.get(pages.xhtmlTestPage);
+    wait.until($ -> "XHTML Test Page".equals(driver.getTitle()));
+  }
+
+  @Test
   public void shouldUseFirefoxOptions() throws InterruptedException {
     DesiredCapabilities caps = new FirefoxOptions()
       .addPreference("browser.startup.page", 1)
@@ -52,8 +54,34 @@ public class MarionetteTest extends JUnit4TestBase {
 
     driver = new FirefoxDriver(caps);
 
-    shortWait.until(
-      input -> "XHTML Test Page".equals(
-        ((JavascriptExecutor) driver).executeScript("return document.title")));
+    wait.until($ -> "XHTML Test Page".equals(driver.getTitle()));
+  }
+
+  @Test
+  public void canSetProfileInFirefoxOptions() throws InterruptedException {
+    FirefoxProfile profile = new FirefoxProfile();
+    profile.setPreference("browser.startup.page", 1);
+    profile.setPreference("browser.startup.homepage", pages.xhtmlTestPage);
+
+    DesiredCapabilities caps = new FirefoxOptions().setProfile(profile)
+        .addTo(DesiredCapabilities.firefox());
+
+    driver = new FirefoxDriver(caps);
+
+    wait.until($ -> "XHTML Test Page".equals(driver.getTitle()));
+  }
+
+  @Test
+  public void canSetProfileInCapabilities() throws InterruptedException {
+    FirefoxProfile profile = new FirefoxProfile();
+    profile.setPreference("browser.startup.page", 1);
+    profile.setPreference("browser.startup.homepage", pages.xhtmlTestPage);
+
+    DesiredCapabilities caps = new DesiredCapabilities();
+    caps.setCapability(FirefoxDriver.PROFILE, profile);
+
+    driver = new FirefoxDriver(caps);
+
+    wait.until($ -> "XHTML Test Page".equals(driver.getTitle()));
   }
 }