+++ rb/lib/selenium/webdriver/firefox/w3c_bridge.rb
@@ -56,9 +56,17 @@ module Selenium
         private
 
         def create_capabilities(opts)
-          caps = opts.delete(:desired_capabilities) || Remote::W3CCapabilities.firefox
+          caps = Remote::W3CCapabilities.firefox
+          caps.merge!(opts.delete(:desired_capabilities)) if opts.key? :desired_capabilities
           firefox_options_caps = caps[:firefox_options] || {}
           caps[:firefox_options] = firefox_options_caps.merge(opts[:firefox_options] || {})
+          if opts.key?(:profile)
+            profile = opts.delete(:profile)
+            unless profile.is_a?(Profile)
+              profile = Profile.from_name(profile)
+            end
+            caps[:firefox_options][:profile] = profile.encoded
+          end
 
           Binary.path = caps[:firefox_options][:binary] if caps[:firefox_options].key?(:binary)
           caps